refactor: flytt brawling-templates og combat-stat-derivasjon til rette lag (#88)
Brawling-maler (punchTemplate, kickTemplate) flyttes fra web-feature til
@glantri/content/equipment. CombatStateCharacterInputs-grensesnitt og
buildCombatStateCharacterInputs-funksjon flyttes til @glantri/rules-engine.
Alle importerende filer oppdatert til ny kildeplassering.
