+++ packages/content/src/equipment/brawlingTemplates.ts
@@ -1,0 +1,144 @@
+import type { WeaponTemplate } from "@glantri/domain";
+
+export const punchTemplate: WeaponTemplate = {
+  id: "virtual-weapon-template-punch",
+  category: "weapon",
+  name: "Punch",
+  subtype: "brawling",
+  tags: ["brawling", "melee", "virtual-workbook"],
+  specificityTypeDefault: "generic",
+  defaultMaterial: "other",
+  baseEncumbrance: 0,
+  baseValue: null,
+  rulesNotes: "Workbook-backed virtual combat row from Themistogenes Weapon1.",
+  roleplayNotes: null,
+  weaponClass: "brawling",
+  weaponSkill: "Brawling",
+  handlingClass: "one_handed",
+  attackModes: [
+    {
+      id: "mode-1",
+      label: "Strike",
+      damageClass: "blunt",
+      ob: 1,
+      obRaw: "1.0",
+      dmb: 0,
+      dmbRaw: "0.0",
+      crit: "AC",
+      armorModifier: "A",
+      provenance: "imported",
+      notes: null,
+    },
+  ],
+  primeAttackType: "Strike",
+  primaryAttackType: "Strike",
+  secondaryAttackType: null,
+  ob1: 1,
+  dmb1: 0,
+  ob2: null,
+  dmb2: null,
+  parry: 0,
+  initiative: 0,
+  range: "1.0",
+  armorMod1: "A",
+  armorMod2: null,
+  crit1: "AC",
+  crit2: null,
+  secondCrit: null,
+  defensiveValue: 0,
+  ammoEncumbrance: null,
+  ammoEncumbranceRaw: null,
+  sourceMetadata: {
+    workbook: "Themistogenes 1.07.xlsx",
+    sheet: "Weapon1",
+    row: 40,
+    sourceRange: "Weapon1!A40:Q40",
+    sourceColumns: {
+      name: "A",
+      skill: "B",
+      primaryAttackLabel: "C",
+      ob1: "D",
+      dmb1: "E",
+      ob2: "F",
+      dmb2: "G",
+      parry: "H",
+      initiative: "I",
+      range: "J",
+      armorMod1: "K",
+      armorMod2: "L",
+      crit1: "M",
+      crit2: "N",
+      encumbrance: "O",
+      defensiveValue: "P",
+      secondCrit: "Q",
+    },
+    rawRow: {
+      A: "Punch",
+      B: "Brawling",
+      C: "Strike",
+      D: "1.0",
+      E: "0.0",
+      F: "",
+      G: "",
+      H: "0.0",
+      I: "0.0",
+      J: "1.0",
+      K: "A",
+      L: "",
+      M: "AC",
+      N: "",
+      O: "0.0",
+      P: "0.0",
+    },
+  },
+  importWarnings: null,
+  durabilityProfile: null,
+};
+
+export const kickTemplate: WeaponTemplate = {
+  ...punchTemplate,
+  id: "virtual-weapon-template-kick",
+  name: "Kick",
+  attackModes: [
+    {
+      id: "mode-1",
+      label: "Strike",
+      damageClass: "blunt",
+      ob: 0,
+      obRaw: "0.0",
+      dmb: 2,
+      dmbRaw: "2.0",
+      crit: "AC",
+      armorModifier: "A",
+      provenance: "imported",
+      notes: null,
+    },
+  ],
+  ob1: 0,
+  dmb1: 2,
+  sourceMetadata: {
+    workbook: "Themistogenes 1.07.xlsx",
+    sheet: "Weapon1",
+    row: 41,
+    sourceRange: "Weapon1!A41:Q41",
+    sourceColumns: punchTemplate.sourceMetadata!.sourceColumns,
+    rawRow: {
+      A: "Kick",
+      B: "Brawling",
+      C: "Strike",
+      D: "0.0",
+      E: "2.0",
+      F: "",
+      G: "",
+      H: "0.0",
+      I: "0.0",
+      J: "1.0",
+      K: "A",
+      L: "",
+      M: "AC",
+      N: "",
+      O: "0.0",
+      P: "0.0",
+    },
+  },
+};
