+++ packages/content/src/equipment/index.ts
@@ -6,6 +6,7 @@ export * from "./themistogenesWeaponEnrichments";
 export * from "./themistogenesWeaponFormulaNormalization";
 export * from "./weaponTemplates";
 export * from "./shieldTemplates";
+export * from "./brawlingTemplates";
 export * from "./systemLocations";
 export * from "./sampleLoadouts";
